ruby libev: add a workaround for signal handling with libev event loop
Currently, we check whether signal is emitted each event loop. The
event loop is broken if signal is emitted with errno=EINTR.

    loop do
      check_signal
      libev_event_loop_run_once # <- It breaks when signal is emitted.
      check_signal
    end

Ruby's signal handler just notifies "signal is emitted". The real
signal handling is done in timer thread. It is other thread. So the
real signal handling may not be occurred immediately.

    loop do
      check_signal
      libev_event_loop_run_once # <- EINTR!
      # If timer thread is ran at this time, the emitted signal can be handled
      # in the below check_signal. In many cases, it is not occurred. :<
      # When timer thread isn't ran, the next libev_event_loop_run_once is ran
      # and blocks this thread. So we can't check signals...
      check_signal
    end

If we can break libev's event loop periodically, we can check emitted
signals periodically. This is the current workaround.