ci package: use Incus instead of Vagrant to fix test job (#215)
Update the runner to `ubuntu-latest` and use Incus instead of Vagrant.

The current test job is broken because `masos-12` runner is not provided
anymore.
Now we can use Incus as a system container that can be used for systemd
tests.
We don't have to use macOS runner and Vagrant anymore.